requirements.txt
ollama
fastapi
uvicorn
jinja2
sqlalchemy
pydantic
tqdm
Pillow
pillow-heif
httpx
python-multipart
watchdog
anthropic
python-dotenv
pymupdf
pytesseract
pytest
pytest-asyncio
faster-whisper
edge-tts
